requirements.txt
annotated-types==0.6.0
anyio==4.3.0
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
exceptiongroup==1.2.0
fastapi==0.110.0
gcloud==0.18.3
google-ai-generativelanguage==0.4.0
google-api-core==2.18.0
google-auth==2.29.0
google-cloud==0.34.0
google-cloud-secret-manager==2.7.2
google-generativeai==0.4.1
googleapis-common-protos==1.63.0
grpc-google-iam-v1==0.12.7
grpcio==1.62.1
grpcio-status==1.62.1
h11==0.14.0
httplib2==0.22.0
idna==3.6
libcst==1.2.0
mypy-extensions==1.0.0
numpy==1.26.4
numpy==1.26.4
oauth2client==4.1.3
opencv-python==4.9.0.80
opencv-python==4.9.0.80
packaging==24.0
proto-plus==1.23.0
protobuf==4.25.3
psycopg2==2.9.9
pyasn1==0.5.1
pyasn1-modules==0.3.0
pydantic==2.6.4
pydantic_core==2.16.3
pyparsing==3.1.2
python-dotenv==1.0.1
python-multipart==0.0.9
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.1
starlette==0.36.3
tqdm==4.66.2
typing-inspect==0.9.0
typing_extensions==4.10.0
urllib3==2.2.1
uvicorn==0.29.0
